Create MultiprocessorMathProb.py
This is a solution that I coded to help me understand this hard math questions: 
A single processor takes 20 milliseconds (ms) to prepare data entries and 0.1n ms to copy the entries, where n is the number of entries. A multiprocessor takes 70 ms to prepare and copy one data entry, and whenever the number of entries is doubled, the amount of time to prepare and copy them increases by 5 ms. Given 120 ms to prepare and copy data entries, which processor type can prepare and copy more entries and how many more entries can it prepare and copy?